Rename mmtk_macro_trace_object to mmtk_macros. Move macros/trace_object to macros/.
