+++ macros/src/plan_trace_object_impl.rs
