+++ macros/src/util.rs
