+++ src/plan/immix/global.rs
@@ -26,7 +26,7 @@ use std::sync::Arc;
 use atomic::Ordering;
 use enum_map::EnumMap;
 
-use mmtk_macro_trace_object::PlanTraceObject;
+use mmtk_macros::PlanTraceObject;
 
 #[derive(PlanTraceObject)]
 pub struct Immix<VM: VMBinding> {
